Change is coming...
These changes are the preliminary to the changes that are about to happen to all the variables that are being used here. ATTENTION: If you want to use this, you will NEED to install Flask, Apache, a SQL GUI, EasyPHP, MySQL.connector, Python (2.7.15 was the version used) and have a knowledge of all of these, including HTML, CSS and possibly Javascript in the future. Right now the Python is very messy. So is the look of the site. But it will stay that way, especially while I rename variables and other assignments. 
Thank you for understanding.
I have no fear that anyone will steal this because it is absolute garbage, but that is okay.
P.S. Something like GitKraken would be good if you want to make additions to the code.